Fixes in command console
Fixed an issue where the font size on a command console input would get smaller as text approached the end of the field.
Fixed an issue where hitting the up/down arrow for the history in the command console would work even if the input field were not focused.
Fixed an issue where old text segments for the large command console would not get deleted.
Fixed an issue where collision for text segments would not update until the scoll view moved.